Removed old Server/PollQueue file + start handling signal
